# health_pal_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api_types.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  models.FoodAnalysisResponse:
    properties:
      analysis_date:
        type: string
      created_at:
        type: string
      detected_items:
        type: string
      food_photo_id:
        type: integer
      id:
        type: integer
      micronutrients_json:
        type: string
      total_calories:
        type: number
      total_carbohydrates:
        type: number
      total_fats:
        type: number
      total_protein:
        type: number
    type: object
  models.FoodPhotoResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      meal_type:
        type: string
      user_id:
        type: integer
    type: object
  models.HPHealthData:
    properties:
      created_at:
        type: string
      data_type:
        description: e.g., "steps", "heart_rate", "sleep"
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      timestamp:
        type: string
      unit:
        description: e.g., "count", "bpm", "hours"
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      value:
        type: number
    type: object
  models.HPHealthDataResponse:
    properties:
      created_at:
        type: string
      data_type:
        type: string
      id:
        type: integer
      timestamp:
        type: string
      unit:
        type: string
      user_id:
        type: integer
      value:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Health Pal backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Health Pal Backend API
  version: "1.0"
paths:
  /api/v1/food-photos/analysis-history:
    get:
      description: Retrieves all food analysis results for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Food analysis history retrieved successfully
          schema:
            items:
              $ref: '#/definitions/models.FoodAnalysisResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get food analysis history
      tags:
      - Food Analysis
  /api/v1/food-photos/analyze:
    post:
      consumes:
      - application/json
      description: Analyzes an uploaded food photo using the Gemini API to extract
        food items and calorie information.
      parameters:
      - description: ID of the food photo to analyze
        in: body
        name: food_photo_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Food photo analyzed successfully
          schema:
            $ref: '#/definitions/models.FoodAnalysisResponse'
        "400":
          description: Invalid request or food photo not found
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Analyze a food photo
      tags:
      - Food Analysis
  /api/v1/food-photos/history:
    get:
      description: Retrieves all food photos uploaded by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Food photo history retrieved successfully
          schema:
            items:
              $ref: '#/definitions/models.FoodPhotoResponse'
            type: array
        "400":
          description: Invalid request or file upload error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get food photo history
      tags:
      - Food Photos
  /api/v1/food-photos/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a food photo for the authenticated user.
      parameters:
      - description: Food photo image file
        in: formData
        name: image
        required: true
        type: file
      - description: Optional description for the food photo
        in: formData
        name: description
        type: string
      - description: Optional meal type (e.g., breakfast, lunch, dinner)
        in: formData
        name: meal_type
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Food photo uploaded successfully
          schema:
            $ref: '#/definitions/models.FoodPhotoResponse'
        "400":
          description: Invalid request or file upload error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload a food photo
      tags:
      - Food Photos
...
  /api/v1/hp-data/history:
    get:
      description: Retrieves historical health data for the authenticated user, with
        optional filters.
      parameters:
      - description: Filter by data type (e.g., 'steps', 'heart_rate')
        in: query
        name: data_type
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Health data history retrieved successfully
          schema:
            items:
              $ref: '#/definitions/models.HPHealthDataResponse'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get HP health data history
      tags:
      - HP Data
  /api/v1/hp-data/upload:
    post:
      consumes:
      - application/json
      description: Uploads health data collected from the user's phone (e.g., steps,
        heart rate).
      parameters:
      - description: Array of health data entries
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/models.HPHealthData'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Health data uploaded successfully
          schema:
            $ref: '#/definitions/api_types.SuccessResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload HP health data
      tags:
      - HP Data
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"